requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==24.4.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.0.7
django-allauth==64.0.0
django-appconf==1.0.6
django-compressor==4.5.1
django-extensions==3.2.3
django-widget-tweaks==1.5.0
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
loguru==0.7.2
msgpack==1.0.8
packaging==24.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
python-decouple==3.8
rcssmin==1.1.2
redis==5.0.8
rjsmin==1.2.2
service-identity==24.1.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
win32-setctime==1.1.0
zope.interface==6.4.post2
